requirements.txt
azure-ai-translation-text==1.0.1
azure-core==1.32.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
isodate==0.7.2
lxml==5.3.0
pillow==11.0.0
python-pptx==1.0.2
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.2.3
XlsxWriter==3.2.0